nit: rename BsonSerializationHelper to not be an extension method on object
